kurang create produk

+++ app/Http/Controllers/Admin/ProdukAdminController.php
@@ -3,9 +3,15 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\KategoriProduk;
 use App\Models\Produk;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
+use Illuminate\Support\Facades\Redirect;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Support\Facades\Storage;
+use Intervention\Image\Drivers\Gd\Driver;
+use Intervention\Image\ImageManager;
 
 class ProdukAdminController extends Controller
 {
@@ -14,7 +20,12 @@ class ProdukAdminController extends Controller
      */
     public function index()
     {
-        return Inertia::render('Admin/Produk/Index');
+        $kategoris = KategoriProduk::all();
+        $produks = Produk::with('kategori')->oldest('created_at')->get();
+        return Inertia::render('Admin/Produk/Index', [
+            'produks' => $produks,
+            'kategoris' => $kategoris
+        ]);
     }
 
     /**
@@ -54,7 +65,53 @@ class ProdukAdminController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        //
+        // dd($request->all());
+        try {
+            $request->validate([
+                'name' => 'required',
+                'produk_ketegori_id' => 'required',
+                'deskripsion' => 'required',
+                'price' => 'required',
+                'image' => 'required',
+            ]);
+
+            $product = Produk::find($id);
+
+            if (!$product) {
+                return back()->with('error', 'Data Produk tidak ditemukan');
+            }
+
+            $product->update([
+                'name' => $request->name,
+                'produk_ketegori_id' => $request->produk_ketegori_id,
+                'deskripsion' => $request->deskripsion,
+                'price' => $request->price,
+            ]);
+
+            if ($request->hasFile('image')) {
+                if ($product->image) {
+                    Storage::disk('public')->delete($product->image);
+                }
+
+                $foto = $request->file('image');
+                $filename = uniqid() . '.webp';
+                $path = 'product-image/' . $filename;
+
+                if (!Storage::disk('public')->exists('product-image')) {
+                    Storage::disk('public')->makeDirectory('product-image');
+                }
+
+                $manager = new ImageManager(new Driver());
+                $image = $manager->read($foto->getRealPath());
+                $image->toWebp()->save(storage_path('app/public/' . $path));
+
+                $product->update(['image' => $path]);
+            }
+
+            return back()->with('success', 'Data Produk berhasil diubah');
+        } catch (\Exception $e) {
+            return back()->with('error', 'Terjadi kesalahan: ' . $e->getMessage());
+        }
     }
 
     /**
@@ -62,7 +119,17 @@ class ProdukAdminController extends Controller
      */
     public function destroy(string $id)
     {
-        //
+        try {
+            $produk = Produk::find($id);
+            if ($produk) {
+                $produk->delete();
+                return back()->with('success', 'data produk berhasil di hapus');
+            } else {
+                return back()->with('error', 'data produk produk tidak ditemukan');
+            }
+        } catch (\Exception $e) {
+            return back()->with('error', $e->getMessage());
+        }
     }
 
     public function restore($id)
@@ -101,7 +168,10 @@ class ProdukAdminController extends Controller
         try {
             $produk = Produk::onlyTrashed()->where('id', $id)->first();
             if ($produk) {
-                $produk->forceDelete();
+                if ($produk->image) {
+                    Storage::disk('public')->delete($produk->image);
+                    $produk->forceDelete();
+                }
                 return back()->with('success', 'Data produk berhasil dihapus permanen');
             } else {
                 return back()->with('error', 'Data produk tidak ditemukan');
